Fix animations for multiple elements
`animations` will now fallback to clone of `DEFAULTS.animations` instead of the object itself to resolve a bug where animations did not display when more than two elements were created.

Font name and size are now stored in one property in the `DEFAULTS` object. They are also now set with one function reducing unnecessary clutter as font properties cannot be edited by the user.

Added a `text` colour in `DEFAULTS` object that utilizes the `darkOrLight` function for determiing text colour.

Renamed `onunfocus` event to `onblur` for clarity and convenience for users that are familiar with web development.

Renamed `accent` colour to `primary` for clarity.

Renamed Essentials library import name to "Essentials". Also fixed `IMPORTED_CORE` variable by renaming to `IMPORTED_ESSENTIALS` as it was changed in the library when migrated to a standalone repository in v0.1.0

Added secondary button for testing with more than one element.

Resolves: #2